Extract ProfileAccessModules from the main Profile component
